+++ src/components/modals/styles/AddContactForm.module.scss
@@ -1,5 +1,5 @@
-@import '../../../assets/styles/global/variables.scss';
-@import '../../../assets/styles/global/mixin.scss';
+@import '../../../../assets/styles/global/variables.scss';
+@import '../../../../assets/styles/global/mixin.scss';
 
 .form {
   width: 318px;
