+++ src/components/modals/styles/Modal.module.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/global/variables.scss';
+@import '../../../../assets/styles/global/variables.scss';
 
 .modal {
   position: fixed;
