+++ src/components/modals/styles/ModalContact.module.scss
@@ -1,5 +1,5 @@
-@import '../../../assets/styles/global/variables.scss';
-@import '../../../assets/styles/global/mixin.scss';
+@import '../../../../assets/styles/global/variables.scss';
+@import '../../../../assets/styles/global/mixin.scss';
 
 .modalContact {
   position: relative;
