+++ src/components/modals/styles/ModalForm.module.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/global/variables.scss';
+@import '../../../../assets/styles/global/variables.scss';
 
 .modalForm {
   position: relative;
