+++ src/components/styles/SectionContactList.module.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/global/variables.scss';
+@import '../../../assets/styles/global/variables.scss';
 
 .contactList {
   margin: 50px auto;
